change the command structure one more time, should not have to do that again. Move cli to seperate file and rework some things so it can run the restart command
