+++ commands/source.js
@@ -3,7 +3,7 @@ module.exports = [
     name: 'source',
     prefix: '!source',
     description: 'Get a link to the source code for this bot',
-    execute: async (message, commands, args) => {
+    execute: async (message, args, context) => {
       const source = `https://github.com/sydney-sisco/magic-8-ball`
       message.reply(source);
     },
