+++ magic-8-ball.js
@@ -12,7 +12,6 @@ const client = new Client({
   ],
 });
 
-
 // load additional commands
 const { loadCommands } = require('./commands/index.js');
 const commands = loadCommands();
@@ -59,10 +58,6 @@ client.on('messageCreate', async message => {
     return;
   }
 
-  if(message.content.startsWith('!restart')) {
-    restart(message);
-  }
-
   // check if message starts with a command prefix
   const matchedCommand = [
     ...commands.values()
@@ -71,7 +66,7 @@ client.on('messageCreate', async message => {
   // if a command is matched, execute it
   if (matchedCommand) {
     const args = message.content.slice(matchedCommand.prefix.length).trim().split(/ +/);
-    matchedCommand.execute(message, commands, args);
+    matchedCommand.execute(message, args, { client, rl, commands });
   }
 
   if (message.content.startsWith(PREFIX)) {
@@ -82,54 +77,5 @@ client.on('messageCreate', async message => {
 
 client.login(process.env.DISCORD_BOT_TOKEN);
 
-
-//
-//// Command line interface
-//
-
-const readline = require('readline');
-
-const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-});
-
-rl.on('line', (input) => {
-    console.log(`Received: ${input}`);
-    if (input === "exit") {
-        rl.close();
-    }
-
-    if (input.startsWith('/say')) {
-      const say = input.substring(5);
-      const channel = client.channels.cache.get(process.env.ADMIN_CHANNEL_ID);
-
-      channel.send(`[Admin]: ${say}`);
-    }
-
-    if (input.startsWith('/restart')) {
-      restart();
-    }
-});
-
-rl.on('close', () => {
-    console.log('Exiting command line interface...');
-    process.exit(0);
-});
-
-const restart = async (message) => {
-
-  const restartMessage = '[System]: Restarting...';
-
-  if (message) {
-    await message.reply(restartMessage);
-  } else {
-
-    const channel = client.channels.cache.get(process.env.ADMIN_CHANNEL_ID);
-    await channel.send(restartMessage);
-  }
-
-  console.log('Restarting bot...');
-  rl.close();
-  client.destroy();
-};
+// load command line interface
+require('./features/command-line-interface.js')(client);
